requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
async-timeout==5.0.1
bcrypt==4.2.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.15
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
joblib==1.4.2
mysql-connector-python==9.1.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.10.5
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.2
redis==4.3.4
redis-client==1.0.1
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uvicorn==0.34.0
wheel==0.44.0
wrapt==1.17.2
